removed compile, added more linting checks and removed jsc styling, some cleanups
