+++ .env
@@ -1,0 +1,8 @@
+### You must Comply! ###
+
+# Slack Stuff
+export CLIENT_ID=1
+export CLIENT_SECRET=2
+
+# Application Stuff
+export PORT=4000
