+++ .eslintrc
@@ -1,53 +1,27 @@
-{
-  "ecmaFeatures": {
-    "modules": true
-  },
-  "env": {
-    "node": true,
-    "es6": true
-  },
-  "rules": {
-    "block-scoped-var": 1,
-    "brace-style": [2, "1tbs", { "allowSingleLine": true }],
-    "consistent-this": 2,
-    "global-strict": 0,
-    "no-div-regex": 2,
-    "no-floating-decimal": 2,
-    "no-self-compare": 2,
-    "no-mixed-spaces-and-tabs": [2, true],
-    "no-nested-ternary": 2,
-    "no-unused-vars": [2, {"vars":"all", "args":"none"}],
-    "radix": 2,
-    "space-after-keywords": [2, "always"],
-    "space-unary-ops": 2,
-    "wrap-iife": [2, "any"],
-    "camelcase": 2,
-    "consistent-return": 1,
-    "max-nested-callbacks": [1, 3],
-    "new-cap": 1,
-    "no-extra-boolean-cast": 1,
-    "no-lonely-if": 1,
-    "no-new": 1,
-    "no-new-wrappers": 1,
-    "no-redeclare": 1,
-    "no-unused-expressions": 1,
-    "no-use-before-define": [1, "nofunc"],
-    "no-warning-comments": 1,
-    "strict": 0,
-    "valid-jsdoc": [1, {
-    "requireReturn": false
-    }],
-    "default-case": 1,
-    "dot-notation": 0,
-    "eol-last": 0,
-    "eqeqeq": 0,
-    "no-trailing-spaces": 0,
-    "no-underscore-dangle": 0,
-    "quotes": 0,
-    "key-spacing": 0,
-    "comma-spacing": 0,
-    "no-multi-spaces": 0,
-    "no-shadow": 0,
-    "no-irregular-whitespace": 0
-  }
+ {
+    "env": {
+        "mocha": true,
+        "node": true,
+        "es6": true       
+    },
+    "rules": {        
+        "camelcase": 0,
+        "eol-last": 1,
+        "handle-callback-err": 1,
+        "new-cap": 0,
+        "no-extra-semi": 1,
+        "no-path-concat": 0,
+        "no-trailing-spaces": 1,
+        "no-undef": 1,
+        "no-underscore-dangle": 0,
+        "no-unused-expressions": 1,
+        "no-unused-vars": 1,
+        "no-use-before-define": [1, "nofunc"],
+        "quotes": [1, "single"],
+        "semi": 2,
+        "space-infix-ops": 1,
+        "strict": 0,
+        "valid-typeof": 1,
+        "indent": [2, 4]
+    }
 }
