+++ .gitignore
@@ -1,5 +1,4 @@
 node_modules
-dist
 .gulp-cache
 npm-debug.log
 coverage
