--- .jscsrc
@@ -1,8 +1,0 @@
-{
-  "esnext": true,
-  "preset": "grunt",
-  "requireCurlyBraces": true,
-  "excludeFiles": ["node_modules/**", "dist", ".gulp-cache", "npm-debug.log", "coverage"],
-  "disallowMixedSpacesAndTabs" : true,
-  "disallowTrailingComma": true
-}
