+++ .nvmrc
@@ -1,1 +1,1 @@
-5.5.0
+5.6.0
