+++ package.json
@@ -4,17 +4,16 @@
   "description": "Musical Chairs for Slack",
   "main": "server.js",
   "scripts": {
-    "test": "gulp test",
-    "start": "node server.js",
-    "lint": "gulp lint",
-    "style": "gulp style",
-    "cover": "gulp cover",
+    "test": "node_modules/.bin/gulp test",
+    "start": "node src/server/server.js",
+    "debug": "node --debug src/server/server.js",
+    "lint": "node_modules/.bin/gulp lint",
+    "cover": "node_modules/.bin/gulp cover",
     "coveralls": "npm run cover && cat ./coverage/lcov.info | coveralls"
   },
   "pre-commit": {
     "run": [
       "lint",
-      "style",
       "test"
     ]
   },
@@ -27,7 +26,7 @@
   "bugs": {
     "url": "https://github.com/mchairs/bot/issues"
   },
-  "homepage": "https://github.com/mchairs/bot#readme",
+  "homepage": "https://mchairs.chat",
   "devDependencies": {
     "babel-core": "^6.4.5",
     "babel-preset-es2015": "^6.3.13",
@@ -39,7 +38,6 @@
     "gulp-eslint": "^1.1.1",
     "gulp-file-cache": "0.0.1",
     "gulp-istanbul": "^0.10.3",
-    "gulp-jscs": "^3.0.2",
     "gulp-mocha": "^2.2.0",
     "gulp-nodemon": "^2.0.6",
     "isparta": "^4.0.0",
