+++ src/server/bot/documents/team.doc.js
@@ -1,10 +1,12 @@
-import mongoose from 'mongoose';
+'use strict';
+
+let mongoose = require('mongoose');
 
 let TeamSchema = mongoose.Schema({
 
 }, {
-  // I am not sure what yet composes a 'Team'
-  strict: false
+    // I am not sure what yet composes a 'Team'
+    strict: false
 });
 
-export default mongoose.model('Team', TeamSchema);
+module.exports = mongoose.model('Team', TeamSchema);
