+++ src/server/bot/services/store.service.js
@@ -1,21 +1,23 @@
-import channel from './channel.service.js';
-import team from './team.service.js';
-import user from './user.service.js';
+'use strict';
 
-export default {
-  teams: {
-    get: team.get,
-    save: team.save,
-    all: team.all
-  },
-  users: {
-    get: user.get,
-    save: user.save,
-    all: user.all
-  },
-  channels: {
-    get: channel.get,
-    save: channel.save,
-    all: channel.all
-  }
+let channel = require('./channel.service.js');
+let team = require('./team.service.js');
+let user = require('./user.service.js');
+
+module.exports = {
+    teams: {
+        get: team.get,
+        save: team.save,
+        all: team.all
+    },
+    users: {
+        get: user.get,
+        save: user.save,
+        all: user.all
+    },
+    channels: {
+        get: channel.get,
+        save: channel.save,
+        all: channel.all
+    }
 };
