+++ src/server/game/player.js
@@ -1,22 +1,22 @@
 'use strict';
 
-export default class Player {
-  constructor(id, chairs) {
-    this.id = id;
-    // Where are you going master?
-    // 'For a drink' :D
-    this.hasTheForce = true;
-    this.chair = null;
-  }
-
-  moveToChair(chair) {
-    if (this.seated()) {
-      this.chair.vacate(this);
+module.exports = class Player {
+    constructor(id) {
+        this.id = id;
+        // Where are you going master?
+        // 'For a drink' :D
+        this.hasTheForce = true;
+        this.chair = null;
     }
-    this.chair = chair;
-  }
 
-  seated() {
-    return !!this.chair;
-  }
+    moveToChair(chair) {
+        if (this.seated()) {
+            this.chair.vacate(this);
+        }
+        this.chair = chair;
+    }
+
+    seated() {
+        return !!this.chair;
+    }
 };
