+++ src/server/log.js
@@ -1,11 +1,13 @@
-import winston from 'winston';
-import moment from 'moment';
+'use strict';
 
-export default new (winston.Logger)({
-  transports: [
-    new (winston.transports.Console)({
-      timestamp: () => `[${moment().format('MMMM Do YYYY, h:mm:ss a')}]`,
-      colorize: true
-    })
-  ]
+let winston = require('winston');
+let moment = require('moment');
+
+module.exports = new (winston.Logger)({
+    transports: [
+        new (winston.transports.Console)({
+            timestamp: () => `[${moment().format('MMMM Do YYYY, h:mm:ss a')}]`,
+            colorize: true
+        })
+    ]
 });
